refactor: split entities.ts and requests.ts into domain-specific folders
The monolithic entities.ts (858 lines) and requests.ts (1369 lines) are
replaced by 16 domain-specific folders under src/types/, each with their
own types.ts, requests.ts, and index.ts barrel. This matches the existing
pattern in src/types/sync/ and makes it easier to navigate types by domain.

No public API changes — the barrel chain still re-exports everything.
