+++ src/todoist-api.ts
@@ -1,78 +1,99 @@
 import {
-    Attachment,
+    DueDateSchema,
+    Task,
+    AddTaskArgs,
+    GetTasksArgs,
+    GetTasksByFilterArgs,
+    UpdateTaskArgs,
+    QuickAddTaskArgs,
+    type MoveTaskArgs,
+    GetCompletedTasksByCompletionDateArgs,
+    GetCompletedTasksByDueDateArgs,
+    GetCompletedTasksResponse,
+    SearchCompletedTasksArgs,
+    GetTasksResponse,
+    GetAllCompletedTasksArgs,
+    GetAllCompletedTasksResponse,
+} from './types/tasks'
+import {
     PersonalProject,
     WorkspaceProject,
-    Label,
+    AddProjectArgs,
+    UpdateProjectArgs,
+    GetProjectsArgs,
+    SearchProjectsArgs,
+    GetProjectCollaboratorsArgs,
+    GetProjectsResponse,
+    GetProjectCollaboratorsResponse,
+    GetArchivedProjectsArgs,
+    GetArchivedProjectsResponse,
+    MoveProjectToWorkspaceArgs,
+    MoveProjectToPersonalArgs,
+    GetArchivedProjectsCountArgs,
+    GetArchivedProjectsCountResponse,
+    GetProjectPermissionsResponse,
+    GetFullProjectArgs,
+    GetFullProjectResponse,
+} from './types/projects'
+import {
     Section,
+    AddSectionArgs,
+    UpdateSectionArgs,
+    GetSectionsArgs,
+    SearchSectionsArgs,
+    GetSectionsResponse,
+} from './types/sections'
+import {
+    Label,
+    AddLabelArgs,
+    UpdateLabelArgs,
+    GetLabelsArgs,
+    SearchLabelsArgs,
+    GetLabelsResponse,
+    GetSharedLabelsArgs,
+    GetSharedLabelsResponse,
+    RenameSharedLabelArgs,
+    RemoveSharedLabelArgs,
+} from './types/labels'
+import {
+    Attachment,
     Comment,
-    Task,
-    CurrentUser,
-    ProductivityStats,
-    DueDateSchema,
+    AddCommentArgs,
+    UpdateCommentArgs,
+    GetProjectCommentsArgs,
+    GetTaskCommentsArgs,
+    GetCommentsResponse,
+} from './types/comments'
+import {
+    REMINDER_DELIVERY_SERVICES,
+    AddReminderArgs,
+    UpdateReminderArgs,
+    AddLocationReminderArgs,
+    UpdateLocationReminderArgs,
+    GetRemindersArgs,
+    GetRemindersResponse,
+    GetLocationRemindersArgs,
+    GetLocationRemindersResponse,
+} from './types/reminders'
+import { CurrentUser } from './types/users'
+import { ProductivityStats } from './types/productivity'
+import { GetActivityLogsArgs, GetActivityLogsResponse } from './types/activity'
+import {
     WorkspaceUser,
     WorkspaceInvitation,
     WorkspacePlanDetails,
     JoinWorkspaceResult,
     Workspace,
-    Backup,
-    IdMapping,
-    MovedId,
-    ProjectActivityStats,
-    ProjectHealth,
-    ProjectHealthContext,
-    ProjectProgress,
-    WorkspaceInsights,
-} from './types/entities'
-import { LOCATION_TRIGGERS } from './types/sync/resources/reminders'
-import {
-    AddCommentArgs,
-    AddLabelArgs,
-    AddLocationReminderArgs,
-    AddProjectArgs,
-    AddReminderArgs,
-    AddSectionArgs,
-    AddTaskArgs,
-    GetProjectCommentsArgs,
-    GetTaskCommentsArgs,
-    GetTasksArgs,
-    GetTasksByFilterArgs,
-    UpdateCommentArgs,
-    UpdateLabelArgs,
-    UpdateLocationReminderArgs,
-    UpdateProjectArgs,
-    UpdateReminderArgs,
-    UpdateSectionArgs,
-    UpdateTaskArgs,
-    QuickAddTaskArgs,
-    GetSharedLabelsArgs,
-    RenameSharedLabelArgs,
-    RemoveSharedLabelArgs,
-    REMINDER_DELIVERY_SERVICES,
-    GetProjectsArgs,
-    SearchProjectsArgs,
-    GetProjectCollaboratorsArgs,
-    GetLabelsArgs,
-    SearchLabelsArgs,
-    GetLabelsResponse,
-    GetTasksResponse,
-    GetProjectsResponse,
-    GetProjectCollaboratorsResponse,
-    GetSectionsArgs,
-    SearchSectionsArgs,
-    GetSectionsResponse,
-    GetSharedLabelsResponse,
-    GetCommentsResponse,
-    type MoveTaskArgs,
-    GetCompletedTasksByCompletionDateArgs,
-    GetCompletedTasksByDueDateArgs,
-    GetCompletedTasksResponse,
-    GetArchivedProjectsArgs,
-    GetArchivedProjectsResponse,
-    SearchCompletedTasksArgs,
-    GetActivityLogsArgs,
-    GetActivityLogsResponse,
-    UploadFileArgs,
-    DeleteUploadArgs,
+    AddWorkspaceArgs,
+    UpdateWorkspaceArgs,
+    GetWorkspaceMembersActivityArgs,
+    GetWorkspaceMembersActivityResponse,
+    GetWorkspaceUserTasksArgs,
+    GetWorkspaceUserTasksResponse,
+    InviteWorkspaceUsersArgs,
+    InviteWorkspaceUsersResponse,
+    UpdateWorkspaceUserArgs,
+    RemoveWorkspaceUserArgs,
     GetWorkspaceInvitationsArgs,
     DeleteWorkspaceInvitationArgs,
     WorkspaceInvitationActionArgs,
@@ -85,31 +106,21 @@ import {
     WorkspaceInvitationsResponse,
     AllWorkspaceInvitationsResponse,
     WorkspaceLogoResponse,
-    MoveProjectToWorkspaceArgs,
-    MoveProjectToPersonalArgs,
-    GetArchivedProjectsCountArgs,
-    GetArchivedProjectsCountResponse,
-    GetProjectPermissionsResponse,
-    GetFullProjectArgs,
-    GetFullProjectResponse,
-    AddWorkspaceArgs,
-    UpdateWorkspaceArgs,
-    GetWorkspaceMembersActivityArgs,
-    GetWorkspaceMembersActivityResponse,
-    GetWorkspaceUserTasksArgs,
-    GetWorkspaceUserTasksResponse,
-    InviteWorkspaceUsersArgs,
-    InviteWorkspaceUsersResponse,
-    UpdateWorkspaceUserArgs,
-    RemoveWorkspaceUserArgs,
+} from './types/workspaces'
+import {
+    ProjectActivityStats,
+    ProjectHealth,
+    ProjectHealthContext,
+    ProjectProgress,
+    WorkspaceInsights,
     GetProjectActivityStatsArgs,
     GetWorkspaceInsightsArgs,
-    GetRemindersArgs,
-    GetRemindersResponse,
-    GetLocationRemindersArgs,
-    GetLocationRemindersResponse,
-    GetAllCompletedTasksArgs,
-    GetAllCompletedTasksResponse,
+} from './types/insights'
+import { Backup, GetBackupsArgs, DownloadBackupArgs } from './types/backups'
+import { IdMapping, MovedId, GetIdMappingsArgs, GetMovedIdsArgs } from './types/id-mappings'
+import { GetOrCreateEmailArgs, GetOrCreateEmailResponse, DisableEmailArgs } from './types/emails'
+import { UploadFileArgs, DeleteUploadArgs } from './types/uploads'
+import {
     ExportTemplateFileArgs,
     ExportTemplateUrlArgs,
     ExportTemplateUrlResponse,
@@ -118,14 +129,8 @@ import {
     ImportTemplateIntoProjectArgs,
     ImportTemplateFromIdArgs,
     ImportTemplateResponse,
-    GetBackupsArgs,
-    DownloadBackupArgs,
-    GetOrCreateEmailArgs,
-    GetOrCreateEmailResponse,
-    DisableEmailArgs,
-    GetIdMappingsArgs,
-    GetMovedIdsArgs,
-} from './types/requests'
+} from './types/templates'
+import { LOCATION_TRIGGERS } from './types/sync/resources/reminders'
 import { CustomFetch, CustomFetchResponse } from './types/http'
 import { request, isSuccess } from './transport/http-client'
 import type { Reminder } from './types'
