+++ src/types/common.ts
@@ -1,0 +1,7 @@
+type SearchArgs = {
+    query: string
+    cursor?: string | null
+    limit?: number
+}
+
+export type { SearchArgs }
