--- src/types/entities.ts
@@ -1,858 +1,0 @@
-import { z } from 'zod'
-import { getProjectUrl, getTaskUrl, getSectionUrl } from '../utils/url-helpers'
-import { hasUncompletablePrefix } from '../utils/uncompletable-helpers'
-
-export const DueDateSchema = z
-    .object({
-        isRecurring: z.boolean(),
-        string: z.string(),
-        date: z.string(),
-    })
-    .extend({
-        datetime: z.string().nullable().optional(),
-        timezone: z.string().nullable().optional(),
-        lang: z.string().nullable().optional(),
-    })
-/**
- * Represents a due date for a task.
- * @see https://developer.todoist.com/api/v1/#tag/Tasks/operation/get_tasks_api_v1_tasks_get
- */
-export type DueDate = z.infer<typeof DueDateSchema>
-
-/** Available duration units for task deadlines. */
-export const DURATION_UNITS = ['minute', 'day'] as const
-/** Unit of time for a task duration. */
-export type DurationUnit = (typeof DURATION_UNITS)[number]
-
-export const DurationSchema = z.object({
-    amount: z.number().positive('Value should be greater than zero'),
-    unit: z.enum(DURATION_UNITS),
-})
-/**
- * Represents a duration for a task deadline.
- * @see https://developer.todoist.com/api/v1/#tag/Tasks
- */
-export type Duration = z.infer<typeof DurationSchema>
-
-export const DeadlineSchema = z.object({
-    date: z.string(),
-    lang: z.string(),
-})
-/**
- * Represents a task deadline.
- */
-export type Deadline = z.infer<typeof DeadlineSchema>
-
-export const TaskSchema = z
-    .object({
-        id: z.string(),
-        userId: z.string(),
-        projectId: z.string(),
-        sectionId: z.string().nullable(),
-        parentId: z.string().nullable(),
-        addedByUid: z.string().nullable(),
-        assignedByUid: z.string().nullable(),
-        responsibleUid: z.string().nullable(),
-        labels: z.array(z.string()),
-        deadline: DeadlineSchema.nullable(),
-        duration: DurationSchema.nullable(),
-        checked: z.boolean(),
-        isDeleted: z.boolean(),
-        addedAt: z.string().nullable(),
-        completedAt: z.string().nullable(),
-        updatedAt: z.string().nullable(),
-        due: DueDateSchema.nullable(),
-        priority: z.number().int(),
-        childOrder: z.number().int(),
-        content: z.string(),
-        description: z.string(),
-        dayOrder: z.number().int(),
-        isCollapsed: z.boolean(),
-        isUncompletable: z.boolean().default(false),
-    })
-    .transform((data) => {
-        // Auto-detect uncompletable status from content prefix
-        const isUncompletable = hasUncompletablePrefix(data.content)
-        return {
-            ...data,
-            isUncompletable,
-            url: getTaskUrl(data.id, data.content),
-        }
-    })
-/**
- * Represents a task in Todoist.
- * @see https://developer.todoist.com/api/v1/#tag/Tasks
- */
-export type Task = z.infer<typeof TaskSchema>
-
-/**
- * Base schema for all project types in Todoist.
- * Contains common fields shared between personal and workspace projects.
- */
-export const BaseProjectSchema = z.object({
-    id: z.string(),
-    canAssignTasks: z.boolean(),
-    childOrder: z.number().int(),
-    color: z.string(),
-    createdAt: z.string().nullable(),
-    isArchived: z.boolean(),
-    isDeleted: z.boolean(),
-    isFavorite: z.boolean(),
-    isFrozen: z.boolean(),
-    name: z.string(),
-    updatedAt: z.string().nullable(),
-    viewStyle: z.string(),
-    defaultOrder: z.number().int(),
-    description: z.string(),
-    isCollapsed: z.boolean(),
-    isShared: z.boolean(),
-})
-
-/**
- * Schema for personal projects in Todoist.
- */
-export const PersonalProjectSchema = BaseProjectSchema.extend({
-    parentId: z.string().nullable(),
-    inboxProject: z.boolean().optional().default(false),
-}).transform((data) => {
-    return {
-        ...data,
-        url: getProjectUrl(data.id, data.name),
-    }
-})
-
-/** Available project visibility levels. */
-export const PROJECT_VISIBILITIES = ['restricted', 'team', 'public'] as const
-/** Visibility level of a workspace project. */
-export type ProjectVisibility = (typeof PROJECT_VISIBILITIES)[number]
-
-export const ProjectVisibilitySchema = z.enum(PROJECT_VISIBILITIES)
-
-/**
- * Schema for workspace projects in Todoist.
- */
-export const WorkspaceProjectSchema = BaseProjectSchema.extend({
-    access: z.object({ visibility: ProjectVisibilitySchema }).optional(),
-    collaboratorRoleDefault: z.string(),
-    folderId: z.string().nullable(),
-    isInviteOnly: z.boolean().nullable(),
-    isLinkSharingEnabled: z.boolean(),
-    role: z.string().nullable(),
-    status: z.string(),
-    workspaceId: z.string(),
-}).transform((data) => {
-    return {
-        ...data,
-        url: getProjectUrl(data.id, data.name),
-    }
-})
-
-/**
- * Represents a personal project in Todoist.
- * @see https://developer.todoist.com/api/v1/#tag/Projects
- */
-export type PersonalProject = z.infer<typeof PersonalProjectSchema>
-
-/**
- * Represents a workspace project in Todoist.
- * @see https://developer.todoist.com/api/v1/#tag/Projects
- */
-export type WorkspaceProject = z.infer<typeof WorkspaceProjectSchema>
-
-/** Available project view styles. */
-export const PROJECT_VIEW_STYLES = ['list', 'board', 'calendar'] as const
-/**
- * View style for a project.
- * @see https://developer.todoist.com/api/v1/#tag/Projects
- */
-export type ProjectViewStyle = (typeof PROJECT_VIEW_STYLES)[number]
-
-export const SectionSchema = z
-    .object({
-        id: z.string(),
-        userId: z.string(),
-        projectId: z.string(),
-        addedAt: z.string(),
-        updatedAt: z.string(),
-        archivedAt: z.string().nullable(),
-        name: z.string(),
-        sectionOrder: z.number().int(),
-        isArchived: z.boolean(),
-        isDeleted: z.boolean(),
-        isCollapsed: z.boolean(),
-    })
-    .transform((data) => {
-        return {
-            ...data,
-            url: getSectionUrl(data.id, data.name),
-        }
-    })
-/**
- * Represents a section in a Todoist project.
- * @see https://developer.todoist.com/api/v1/#tag/Sections
- */
-export type Section = z.infer<typeof SectionSchema>
-
-export const LabelSchema = z.object({
-    id: z.string(),
-    order: z.number().int().nullable(),
-    name: z.string(),
-    color: z.string(),
-    isFavorite: z.boolean(),
-})
-/**
- * Represents a label in Todoist.
- * @see https://developer.todoist.com/api/v1/#tag/Labels
- */
-export type Label = z.infer<typeof LabelSchema>
-
-/** Available file attachment upload states. */
-export const UPLOAD_STATES = ['pending', 'completed'] as const
-/** Upload state of a file attachment. */
-export type UploadState = (typeof UPLOAD_STATES)[number]
-
-export const AttachmentSchema = z
-    .object({
-        resourceType: z.string(),
-    })
-    .extend({
-        fileName: z.string().nullable().optional(),
-        fileSize: z.number().int().nullable().optional(),
-        fileType: z.string().nullable().optional(),
-        fileUrl: z.string().nullable().optional(),
-        fileDuration: z.number().int().nullable().optional(),
-        uploadState: z.enum(UPLOAD_STATES).nullable().optional(),
-        image: z.string().nullable().optional(),
-        imageWidth: z.number().int().nullable().optional(),
-        imageHeight: z.number().int().nullable().optional(),
-        url: z.string().nullable().optional(),
-        title: z.string().nullable().optional(),
-    })
-/**
- * Represents a file attachment in a comment.
- * @see https://developer.todoist.com/api/v1/#tag/Sync/Comments/File-Attachments
- */
-export type Attachment = z.infer<typeof AttachmentSchema>
-
-export const RawCommentSchema = z
-    .object({
-        id: z.string(),
-        itemId: z.string().optional(),
-        projectId: z.string().optional(),
-        content: z.string(),
-        postedAt: z.string(),
-        fileAttachment: AttachmentSchema.nullable(),
-        postedUid: z.string(),
-        uidsToNotify: z.array(z.string()).nullable(),
-        reactions: z.record(z.string(), z.array(z.string())).nullable(),
-        isDeleted: z.boolean(),
-    })
-    .refine(
-        (data) => {
-            // At least one of itemId or projectId must be present
-            return (
-                (data.itemId !== undefined && data.projectId === undefined) ||
-                (data.itemId === undefined && data.projectId !== undefined)
-            )
-        },
-        {
-            message: 'Exactly one of itemId or projectId must be provided',
-        },
-    )
-
-/**
- * Represents a raw comment response from the API.
- * @see https://developer.todoist.com/api/v1/#tag/Comments
- */
-export type RawComment = z.infer<typeof RawCommentSchema>
-
-export const CommentSchema = RawCommentSchema.transform((data) => {
-    const { itemId, ...rest } = data
-    return {
-        ...rest,
-        // Map itemId to taskId for backwards compatibility
-        taskId: itemId,
-    }
-})
-/**
- * Represents a comment in Todoist.
- * @see https://developer.todoist.com/api/v1/#tag/Comments
- */
-export type Comment = z.infer<typeof CommentSchema>
-
-export const UserSchema = z.object({
-    id: z.string(),
-    name: z.string(),
-    email: z.string(),
-})
-/**
- * Represents a user in Todoist (simplified for collaborators).
- * @see https://developer.todoist.com/api/v1/#tag/User
- */
-export type User = z.infer<typeof UserSchema>
-
-export const TimezoneInfoSchema = z.object({
-    gmtString: z.string(),
-    hours: z.number().int(),
-    isDst: z.number().int(),
-    minutes: z.number().int(),
-    timezone: z.string(),
-})
-
-/** Available user premium statuses. */
-export const PREMIUM_STATUSES = [
-    'not_premium',
-    'current_personal_plan',
-    'legacy_personal_plan',
-    'teams_business_member',
-] as const
-/** Premium subscription status of a user. */
-export type PremiumStatus = (typeof PREMIUM_STATUSES)[number]
-
-export const CurrentUserSchema = z.object({
-    id: z.string(),
-    email: z.string(),
-    fullName: z.string(),
-    avatarBig: z.string().nullish(),
-    avatarMedium: z.string().nullish(),
-    avatarS640: z.string().nullish(),
-    avatarSmall: z.string().nullish(),
-    businessAccountId: z.string().nullable(),
-    isPremium: z.boolean(),
-    premiumStatus: z.enum(PREMIUM_STATUSES),
-    dateFormat: z.number().int(),
-    timeFormat: z.number().int(),
-    weeklyGoal: z.number().int(),
-    dailyGoal: z.number().int(),
-    completedCount: z.number().int(),
-    completedToday: z.number().int(),
-    karma: z.number(),
-    karmaTrend: z.string(),
-    lang: z.string(),
-    nextWeek: z.number().int(),
-    startDay: z.number().int(),
-    startPage: z.string(),
-    tzInfo: TimezoneInfoSchema,
-    inboxProjectId: z.string(),
-    daysOff: z.array(z.number().int()),
-    weekendStartDay: z.number().int(),
-})
-/**
- * Represents the current authenticated user with detailed information.
- * @see https://developer.todoist.com/api/v1/#tag/User
- */
-export type CurrentUser = z.infer<typeof CurrentUserSchema>
-
-export const StreakSchema = z.object({
-    count: z.number(),
-    start: z.string(),
-    end: z.string(),
-})
-
-export const CompletedItemSchema = z.object({
-    id: z.string(),
-    completed: z.number(),
-})
-
-export const ItemsWithDateSchema = z.object({
-    items: z.array(CompletedItemSchema),
-    totalCompleted: z.number(),
-})
-
-export const KarmaUpdateSchema = z.object({
-    time: z.string(),
-    newKarma: z.number(),
-    positiveKarma: z.number(),
-    negativeKarma: z.number(),
-    positiveKarmaReasons: z.array(z.any()),
-    negativeKarmaReasons: z.array(z.any()),
-})
-
-export const ProductivityStatsSchema = z.object({
-    completedCount: z.number(),
-    daysItems: z.array(
-        ItemsWithDateSchema.extend({
-            date: z.string(),
-        }),
-    ),
-    goals: z.object({
-        currentDailyStreak: StreakSchema,
-        currentWeeklyStreak: StreakSchema,
-        dailyGoal: z.number(),
-        ignoreDays: z.array(z.number()),
-        karmaDisabled: z.number(),
-        lastDailyStreak: StreakSchema,
-        lastWeeklyStreak: StreakSchema,
-        maxDailyStreak: StreakSchema,
-        maxWeeklyStreak: StreakSchema,
-        user: z.string(),
-        userId: z.string(),
-        vacationMode: z.number(),
-        weeklyGoal: z.number(),
-    }),
-    karma: z.number(),
-    karmaGraphData: z.array(
-        z.object({
-            date: z.string(),
-            karmaAvg: z.number(),
-        }),
-    ),
-    karmaLastUpdate: z.number(),
-    karmaTrend: z.string(),
-    karmaUpdateReasons: z.array(KarmaUpdateSchema),
-    projectColors: z.record(z.string(), z.string()),
-    weekItems: z.array(
-        ItemsWithDateSchema.extend({
-            from: z.string(),
-            to: z.string(),
-        }),
-    ),
-})
-
-/**
- * Represents the Productivity stats for the authenticated user.
- * @see https://developer.todoist.com/api/v1/#tag/User/operation/get_productivity_stats_api_v1_tasks_completed_stats_get
- */
-export type ProductivityStats = z.infer<typeof ProductivityStatsSchema>
-
-export const ColorSchema = z.object({
-    /** The key of the color (i.e. 'berry_red') */
-    key: z.string(),
-    /** The display name of the color (i.e. 'Berry Red') */
-    displayName: z.string(),
-    /** The hex value of the color (i.e. '#b8255f') */
-    hexValue: z.string(),
-})
-/**
- * Represents a color in Todoist.
- * @see https://developer.todoist.com/api/v1/#tag/Colors
- */
-export type Color = z.infer<typeof ColorSchema>
-
-/**
- * Type hints for known object types. Accepts any string for forward compatibility.
- */
-export type ActivityObjectType = 'task' | 'comment' | 'project' | (string & Record<string, never>)
-
-/**
- * Type hints for known event types. Accepts any string for forward compatibility.
- */
-export type ActivityEventType =
-    | 'added'
-    | 'updated'
-    | 'deleted'
-    | 'completed'
-    | 'uncompleted'
-    | 'archived'
-    | 'unarchived'
-    | 'shared'
-    | 'left'
-    | (string & Record<string, never>)
-
-// `string & Record<string, never>` is a TypeScript trick that allows any string
-// to be passed (for forward compatibility) while still providing autocomplete
-// for the known literal values above.
-type ModernActivityObjectType = 'task' | 'comment' | 'project' | (string & Record<string, never>)
-
-/**
- * Combined object:event filter string for the `objectEventTypes` parameter of `getActivityLogs`.
- *
- * Uses modern object type names (`task`, `comment`) rather than legacy API names (`item`, `note`).
- * Either part may be omitted:
- * - `'task:added'` — task additions only
- * - `'task:'` — all events for tasks
- * - `':deleted'` — all deleted events across all object types
- *
- * The final `string & Record<string, never>` member allows any arbitrary string
- * to be passed for forward compatibility, while still providing autocomplete for
- * the known combinations above.
- */
-export type ActivityObjectEventType =
-    | `${ModernActivityObjectType}:${ActivityEventType}`
-    | `${ModernActivityObjectType}:`
-    | `:${ActivityEventType}`
-    | (string & Record<string, never>)
-
-/**
- * Flexible object containing event-specific data.
- * Uses z.record to accept any properties for forward compatibility.
- */
-export const ActivityEventExtraDataSchema = z.record(z.string(), z.any()).nullable()
-export type ActivityEventExtraData = z.infer<typeof ActivityEventExtraDataSchema>
-
-/**
- * Activity log event schema. Accepts unknown fields for forward compatibility.
- */
-export const ActivityEventSchema = z
-    .object({
-        objectType: z.string(),
-        objectId: z.string(),
-        eventType: z.string(),
-        eventDate: z.string(),
-        id: z
-            .union([z.string(), z.number()])
-            .transform((val) => val?.toString() ?? null)
-            .nullable(),
-        parentProjectId: z.string().nullable(),
-        parentItemId: z.string().nullable(),
-        initiatorId: z.string().nullable(),
-        extraData: ActivityEventExtraDataSchema,
-    })
-    .catchall(z.any())
-
-/**
- * Represents an activity log event in Todoist.
- */
-export type ActivityEvent = z.infer<typeof ActivityEventSchema>
-
-/** Available project collaborator roles. */
-export const COLLABORATOR_ROLES = [
-    'CREATOR',
-    'ADMIN',
-    'READ_WRITE',
-    'EDIT_ONLY',
-    'COMPLETE_ONLY',
-] as const
-/** Role of a collaborator in a project. */
-export type CollaboratorRole = (typeof COLLABORATOR_ROLES)[number]
-
-/**
- * Available workspace roles.
- */
-export const WORKSPACE_ROLES = ['ADMIN', 'MEMBER', 'GUEST'] as const
-
-/**
- * Role of a user within a workspace.
- */
-export type WorkspaceRole = (typeof WORKSPACE_ROLES)[number]
-
-export const WorkspaceRoleSchema = z.enum(WORKSPACE_ROLES)
-
-export const WorkspaceUserSchema = z.object({
-    userId: z.string(),
-    workspaceId: z.string(),
-    userEmail: z.string(),
-    fullName: z.string(),
-    timezone: z.string(),
-    role: WorkspaceRoleSchema,
-    imageId: z.string().nullable(),
-    isDeleted: z.boolean().default(false),
-})
-
-/**
- * Represents a user within a workspace (MemberView from API).
- */
-export type WorkspaceUser = z.infer<typeof WorkspaceUserSchema>
-
-export const WorkspaceInvitationSchema = z.object({
-    id: z.string().default('0'),
-    inviterId: z.string(),
-    userEmail: z.string(),
-    workspaceId: z.string(),
-    role: WorkspaceRoleSchema,
-    isExistingUser: z.boolean(),
-})
-
-/**
- * Represents a workspace invitation.
- */
-export type WorkspaceInvitation = z.infer<typeof WorkspaceInvitationSchema>
-
-export const PlanPriceSchema = z.object({
-    currency: z.string(),
-    amount: z.union([z.number(), z.string()]),
-    interval: z.string().optional(),
-})
-
-/**
- * Plan pricing information.
- */
-export type PlanPrice = z.infer<typeof PlanPriceSchema>
-
-export const FormattedPriceListingSchema = z.object({
-    currency: z.string().optional(),
-    amount: z.number().optional(),
-    interval: z.string().optional(),
-    formatted: z.string().optional(),
-})
-
-/**
- * Formatted price listing for workspace plans.
- */
-export type FormattedPriceListing = z.infer<typeof FormattedPriceListingSchema>
-
-/** Available workspace plan names. */
-export const WORKSPACE_CURRENT_PLANS = ['Business', 'Starter'] as const
-/** Display name of a workspace plan. */
-export type WorkspaceCurrentPlan = (typeof WORKSPACE_CURRENT_PLANS)[number]
-
-/** Available workspace plan statuses. */
-export const WORKSPACE_PLAN_STATUSES = [
-    'Active',
-    'Downgraded',
-    'Cancelled',
-    'NeverSubscribed',
-] as const
-/** Subscription status of a workspace plan. */
-export type WorkspacePlanStatus = (typeof WORKSPACE_PLAN_STATUSES)[number]
-
-export const WorkspacePlanDetailsSchema = z.object({
-    currentMemberCount: z.number(),
-    currentPlan: z.enum(WORKSPACE_CURRENT_PLANS),
-    currentPlanStatus: z.enum(WORKSPACE_PLAN_STATUSES),
-    downgradeAt: z.string().nullable(),
-    currentActiveProjects: z.number(),
-    maximumActiveProjects: z.number(),
-    priceList: z.array(FormattedPriceListingSchema),
-    workspaceId: z.number(),
-    isTrialing: z.boolean(),
-    trialEndsAt: z.string().nullable(),
-    cancelAtPeriodEnd: z.boolean(),
-    hasTrialed: z.boolean(),
-    planPrice: PlanPriceSchema.nullable(),
-    hasBillingPortal: z.boolean(),
-    hasBillingPortalSwitchToAnnual: z.boolean(),
-})
-
-/**
- * Represents workspace plan and billing details.
- */
-export type WorkspacePlanDetails = z.infer<typeof WorkspacePlanDetailsSchema>
-
-export const JoinWorkspaceResultSchema = z.object({
-    customSortingApplied: z.boolean(),
-    projectSortPreference: z.string(),
-    role: WorkspaceRoleSchema,
-    userId: z.string(),
-    workspaceId: z.string(),
-})
-
-/**
- * Result returned when successfully joining a workspace.
- */
-export type JoinWorkspaceResult = z.infer<typeof JoinWorkspaceResultSchema>
-
-/**
- * Available workspace plans.
- */
-export const WORKSPACE_PLANS = ['STARTER', 'BUSINESS'] as const
-
-/**
- * Workspace plan type.
- */
-export type WorkspacePlan = (typeof WORKSPACE_PLANS)[number]
-
-export const WorkspacePlanSchema = z.enum(WORKSPACE_PLANS)
-
-/**
- * Workspace resource limits.
- */
-export const WorkspaceLimitsSchema = z
-    .object({
-        current: z.record(z.string(), z.any()).nullable(),
-        next: z.record(z.string(), z.any()).nullable(),
-    })
-    .catchall(z.any())
-
-export type WorkspaceLimits = z.infer<typeof WorkspaceLimitsSchema>
-
-/**
- * Workspace properties (flexible object for unknown fields).
- */
-export const WorkspacePropertiesSchema = z.record(z.string(), z.unknown())
-export type WorkspaceProperties = z.infer<typeof WorkspacePropertiesSchema>
-
-/**
- * Represents a workspace in Todoist.
- */
-export const WorkspaceSchema = z.object({
-    id: z.string(),
-    name: z.string(),
-    plan: WorkspacePlanSchema,
-    role: WorkspaceRoleSchema,
-    inviteCode: z.string(),
-    isLinkSharingEnabled: z.boolean(),
-    isGuestAllowed: z.boolean(),
-    limits: WorkspaceLimitsSchema,
-    logoBig: z.string().nullish(),
-    logoMedium: z.string().nullish(),
-    logoSmall: z.string().nullish(),
-    logoS640: z.string().nullish(),
-    createdAt: z.string(),
-    creatorId: z.string(),
-    properties: WorkspacePropertiesSchema,
-})
-
-export type Workspace = z.infer<typeof WorkspaceSchema>
-
-export const MemberActivityInfoSchema = z.object({
-    userId: z.string(),
-    tasksAssigned: z.number().int(),
-    tasksOverdue: z.number().int(),
-})
-/**
- * Represents activity information for a workspace member.
- */
-export type MemberActivityInfo = z.infer<typeof MemberActivityInfoSchema>
-
-export const WorkspaceUserTaskSchema = z.object({
-    id: z.string(),
-    content: z.string(),
-    responsibleUid: z.string().nullable(),
-    due: DueDateSchema.nullable(),
-    deadline: DeadlineSchema.nullable(),
-    labels: z.array(z.string()),
-    notesCount: z.number().int(),
-    projectId: z.string(),
-    projectName: z.string(),
-    priority: z.number().int(),
-    description: z.string(),
-    isOverdue: z.boolean(),
-})
-/**
- * Represents a task assigned to a workspace user.
- */
-export type WorkspaceUserTask = z.infer<typeof WorkspaceUserTaskSchema>
-
-// Insights entities
-
-export const DayActivitySchema = z.object({
-    date: z.string(),
-    totalCount: z.number().int(),
-})
-/** Daily activity count for a specific date. */
-export type DayActivity = z.infer<typeof DayActivitySchema>
-
-export const WeekRollupSchema = z.object({
-    fromDate: z.string(),
-    toDate: z.string(),
-    totalCount: z.number().int(),
-})
-/** Weekly rollup of activity data. */
-export type WeekRollup = z.infer<typeof WeekRollupSchema>
-
-export const ProjectActivityStatsSchema = z.object({
-    dayItems: z.array(DayActivitySchema),
-    weekItems: z.array(WeekRollupSchema).nullable(),
-})
-/** Project activity statistics with daily and optional weekly rollups. */
-export type ProjectActivityStats = z.infer<typeof ProjectActivityStatsSchema>
-
-/** Available project health statuses. */
-export const HEALTH_STATUSES = [
-    'UNKNOWN',
-    'ON_TRACK',
-    'AT_RISK',
-    'CRITICAL',
-    'EXCELLENT',
-    'ERROR',
-] as const
-/** Health status of a project. */
-export type HealthStatus = (typeof HEALTH_STATUSES)[number]
-
-export const TaskRecommendationSchema = z.object({
-    taskId: z.string(),
-    recommendation: z.string(),
-})
-/** A recommendation for a specific task. */
-export type TaskRecommendation = z.infer<typeof TaskRecommendationSchema>
-
-export const ProjectHealthSchema = z.object({
-    status: z.enum(HEALTH_STATUSES),
-    description: z.string().nullable().optional(),
-    descriptionSummary: z.string().nullable().optional(),
-    taskRecommendations: z.array(TaskRecommendationSchema).nullable().optional(),
-    projectId: z.string().nullable().optional(),
-    updatedAt: z.string().nullable().optional(),
-    isStale: z.boolean().default(false),
-    updateInProgress: z.boolean().default(false),
-})
-/** Project health status and recommendations. */
-export type ProjectHealth = z.infer<typeof ProjectHealthSchema>
-
-export const ProjectMetricsSchema = z.object({
-    totalTasks: z.number().int(),
-    completedTasks: z.number().int(),
-    overdueTasks: z.number().int(),
-    tasksCreatedThisWeek: z.number().int(),
-    tasksCompletedThisWeek: z.number().int(),
-    averageCompletionTime: z.number().nullable(),
-})
-/** Project metrics summary. */
-export type ProjectMetrics = z.infer<typeof ProjectMetricsSchema>
-
-export const TaskContextSchema = z.object({
-    id: z.string(),
-    content: z.string(),
-    due: z.string().nullable().optional(),
-    deadline: z.string().nullable().optional(),
-    priority: z.string(),
-    isCompleted: z.boolean(),
-    createdAt: z.string(),
-    updatedAt: z.string(),
-    completedAt: z.string().nullable(),
-    completedByUid: z.string().nullable(),
-    labels: z.array(z.string()),
-})
-/** Task context used in health analysis. */
-export type TaskContext = z.infer<typeof TaskContextSchema>
-
-export const ProjectHealthContextSchema = z.object({
-    projectId: z.string(),
-    projectName: z.string(),
-    projectDescription: z.string().nullable(),
-    projectMetrics: ProjectMetricsSchema,
-    tasks: z.array(TaskContextSchema),
-    language: z.string().nullable().optional(),
-})
-/** Full project context for health analysis. */
-export type ProjectHealthContext = z.infer<typeof ProjectHealthContextSchema>
-
-export const ProjectProgressSchema = z.object({
-    projectId: z.string(),
-    completedCount: z.number().int(),
-    activeCount: z.number().int(),
-    progressPercent: z.number().int(),
-})
-/** Project progress with completion counts and percentage. */
-export type ProjectProgress = z.infer<typeof ProjectProgressSchema>
-
-export const ProjectInsightSchema = z.object({
-    projectId: z.string(),
-    health: ProjectHealthSchema.nullable(),
-    progress: ProjectProgressSchema.nullable(),
-})
-/** Insight data for a single project within a workspace. */
-export type ProjectInsight = z.infer<typeof ProjectInsightSchema>
-
-export const WorkspaceInsightsSchema = z.object({
-    folderId: z.string().nullable(),
-    projectInsights: z.array(ProjectInsightSchema),
-})
-/** Workspace insights grouped by folder. */
-export type WorkspaceInsights = z.infer<typeof WorkspaceInsightsSchema>
-
-// Backups
-
-export const BackupSchema = z.object({
-    version: z.string(),
-    url: z.string(),
-})
-/** Represents a backup snapshot. */
-export type Backup = z.infer<typeof BackupSchema>
-
-// ID Mappings
-
-export const IdMappingSchema = z.object({
-    oldId: z.string().nullable(),
-    newId: z.string().nullable(),
-})
-/** Mapping between old and new IDs after migration. */
-export type IdMapping = z.infer<typeof IdMappingSchema>
-
-export const MovedIdSchema = z.object({
-    oldId: z.string(),
-    newId: z.string(),
-})
-/** A moved ID pair (old to new). */
-export type MovedId = z.infer<typeof MovedIdSchema>
