--- src/types/requests.ts
@@ -1,1368 +1,0 @@
-import type { RequireAllOrNone, RequireOneOrNone, RequireExactlyOne } from 'type-fest'
-import type { ColorKey } from '../utils/colors'
-import type {
-    ActivityEvent,
-    ActivityObjectEventType,
-    CollaboratorRole,
-    Comment,
-    DueDate,
-    Duration,
-    Label,
-    MemberActivityInfo,
-    PersonalProject,
-    ProjectViewStyle,
-    ProjectVisibility,
-    Section,
-    Task,
-    User,
-    WorkspaceInvitation,
-    WorkspaceProject,
-    WorkspaceRole,
-    WorkspaceUser,
-    WorkspaceUserTask,
-} from './entities'
-import type { LocationTrigger, Reminder, LocationReminder } from './sync/resources/reminders'
-
-/**
- * Arguments for creating a new task.
- * @see https://developer.todoist.com/api/v1/#tag/Tasks/operation/create_task_api_v1_tasks_post
- */
-export type AddTaskArgs = {
-    content: string
-    description?: string
-    projectId?: string
-    sectionId?: string
-    parentId?: string
-    order?: number
-    labels?: string[]
-    priority?: number
-    assigneeId?: string
-    dueString?: string
-    dueLang?: string
-    deadlineLang?: string
-    deadlineDate?: string
-    isUncompletable?: boolean
-} & RequireOneOrNone<{
-    dueDate?: string
-    dueDatetime?: string
-}> &
-    RequireAllOrNone<{
-        duration?: Duration['amount']
-        durationUnit?: Duration['unit']
-    }>
-
-/**
- * Arguments for retrieving tasks.
- * @see https://developer.todoist.com/api/v1/#tag/Tasks/operation/get_tasks_api_v1_tasks_get
- */
-export type GetTasksArgs = {
-    projectId?: string
-    sectionId?: string
-    parentId?: string
-    label?: string
-    ids?: string[]
-    cursor?: string | null
-    limit?: number
-}
-
-/**
- * Arguments for retrieving tasks by filter.
- * @see https://developer.todoist.com/api/v1/#tag/Tasks/operation/get_tasks_by_filter_api_v1_tasks_filter_get
- */
-export type GetTasksByFilterArgs = {
-    query: string
-    lang?: string
-    cursor?: string | null
-    limit?: number
-}
-
-/**
- * Arguments for retrieving completed tasks by completion date.
- * @see https://developer.todoist.com/api/v1/#tag/Tasks/operation/tasks_completed_by_completion_date_api_v1_tasks_completed_by_completion_date_get
- */
-export type GetCompletedTasksByCompletionDateArgs = {
-    since: string
-    until: string
-    workspaceId?: string | null
-    projectId?: string | null
-    sectionId?: string | null
-    parentId?: string | null
-    filterQuery?: string | null
-    filterLang?: string | null
-    cursor?: string | null
-    limit?: number
-    publicKey?: string | null
-}
-
-/**
- * Arguments for retrieving completed tasks by due date.
- * @see https://developer.todoist.com/api/v1/#tag/Tasks/operation/tasks_completed_by_due_date_api_v1_tasks_completed_by_due_date_get
- */
-export type GetCompletedTasksByDueDateArgs = {
-    since: string
-    until: string
-    workspaceId?: string | null
-    projectId?: string | null
-    sectionId?: string | null
-    parentId?: string | null
-    filterQuery?: string | null
-    filterLang?: string | null
-    cursor?: string | null
-    limit?: number
-}
-
-/**
- * Arguments for searching completed tasks.
- */
-export type SearchCompletedTasksArgs = {
-    query: string
-    cursor?: string | null
-    limit?: number
-}
-
-/**
- * @see https://developer.todoist.com/api/v1/#tag/Tasks/operation/get_tasks_api_v1_tasks_get
- */
-export type GetTasksResponse = {
-    results: Task[]
-    nextCursor: string | null
-}
-
-/**
- * @see https://developer.todoist.com/api/v1/#tag/Tasks/operation/tasks_completed_by_due_date_api_v1_tasks_completed_by_due_date_get
- * @see https://developer.todoist.com/api/v1/#tag/Tasks/operation/tasks_completed_by_completion_date_api_v1_tasks_completed_by_completion_date_get
- */
-export type GetCompletedTasksResponse = {
-    items: Task[]
-    nextCursor: string | null
-}
-
-/**
- * Arguments for updating a task.
- * @see https://developer.todoist.com/api/v1/#tag/Tasks/operation/update_task_api_v1_tasks__task_id__post
- */
-export type UpdateTaskArgs = {
-    content?: string
-    description?: string
-    labels?: string[]
-    priority?: number
-    /**
-     * Sort order of the task within its parent/project.
-     * Internally mapped to `child_order` in the API payload.
-     */
-    order?: number
-    /**
-     * Natural language due date.
-     * Use `"no date"` to clear a due date, or `null` as an SDK alias for the same behavior.
-     */
-    dueString?: (string & {}) | 'no date' | null
-    dueLang?: string | null
-    assigneeId?: string | null
-    /**
-     * Deadline date in `YYYY-MM-DD` format.
-     * Use `null` to clear a task deadline.
-     */
-    deadlineDate?: string | null
-    deadlineLang?: string | null
-    isUncompletable?: boolean
-} & RequireOneOrNone<{
-    dueDate?: string
-    dueDatetime?: string
-}> &
-    RequireAllOrNone<{
-        duration?: Duration['amount']
-        durationUnit?: Duration['unit']
-    }>
-
-/**
- * Arguments for quick adding a task.
- * @see https://developer.todoist.com/api/v1/#tag/Tasks/operation/quick_add_api_v1_tasks_quick_post
- */
-export type QuickAddTaskArgs = {
-    text: string
-    note?: string
-    reminder?: string
-    autoReminder?: boolean
-    meta?: boolean
-    isUncompletable?: boolean
-}
-
-/**
- * Response from quick adding a task.
- * @see https://developer.todoist.com/api/v1/#tag/Tasks/operation/quick_add_api_v1_tasks_quick_post
- */
-
-/**
- * Arguments for moving a task.
- * @see https://developer.todoist.com/api/v1/#tag/Tasks/operation/move_task_api_v1_tasks__task_id__move_post
- */
-export type MoveTaskArgs = RequireExactlyOne<{
-    projectId?: string
-    sectionId?: string
-    parentId?: string
-}>
-
-/**
- * Arguments for retrieving projects.
- * @see https://developer.todoist.com/api/v1/#tag/Projects/operation/get_projects_api_v1_projects_get
- */
-export type GetProjectsArgs = {
-    cursor?: string | null
-    limit?: number
-}
-
-/**
- * Arguments for searching projects.
- */
-type SearchArgs = {
-    query: string
-    cursor?: string | null
-    limit?: number
-}
-export type SearchProjectsArgs = SearchArgs
-
-/**
- * Response from retrieving projects.
- * @see https://developer.todoist.com/api/v1/#tag/Projects/operation/get_projects_api_v1_projects_get
- */
-export type GetProjectsResponse = {
-    results: (PersonalProject | WorkspaceProject)[]
-    nextCursor: string | null
-}
-
-/**
- * Arguments for retrieving archived projects.
- * @see https://developer.todoist.com/api/v1/#tag/Projects/operation/get_archived_projects_api_v1_projects_archived_get
- */
-export type GetArchivedProjectsArgs = {
-    cursor?: string | null
-    limit?: number
-}
-
-/**
- * Response from retrieving archived projects.
- * @see https://developer.todoist.com/api/v1/#tag/Projects/operation/get_archived_projects_api_v1_projects_archived_get
- */
-export type GetArchivedProjectsResponse = {
-    results: (PersonalProject | WorkspaceProject)[]
-    nextCursor: string | null
-}
-
-/**
- * Arguments for creating a new project.
- * @see https://developer.todoist.com/api/v1/#tag/Projects/operation/create_project_api_v1_projects_post
- */
-export type AddProjectArgs = {
-    name: string
-    parentId?: string
-    color?: ColorKey
-    isFavorite?: boolean
-    viewStyle?: ProjectViewStyle
-    workspaceId?: string
-}
-
-/**
- * Arguments for updating a project.
- * @see https://developer.todoist.com/api/v1/#tag/Projects/operation/update_project_api_v1_projects__project_id__post
- */
-export type UpdateProjectArgs = {
-    name?: string
-    color?: ColorKey
-    isFavorite?: boolean
-    viewStyle?: ProjectViewStyle
-}
-
-/**
- * Arguments for retrieving project collaborators.
- * @see https://developer.todoist.com/api/v1/#tag/Projects/operation/get_project_collaborators_api_v1_projects__project_id__collaborators_get
- */
-export type GetProjectCollaboratorsArgs = {
-    cursor?: string | null
-    limit?: number
-}
-
-/**
- * Response from retrieving project collaborators.
- * @see https://developer.todoist.com/api/v1/#tag/Projects/operation/get_project_collaborators_api_v1_projects__project_id__collaborators_get
- */
-export type GetProjectCollaboratorsResponse = {
-    results: User[]
-    nextCursor: string | null
-}
-
-/**
- * Arguments for retrieving sections.
- * @see https://developer.todoist.com/api/v1/#tag/Sections/operation/get_sections_api_v1_sections_get
- */
-export type GetSectionsArgs = {
-    projectId?: string | null
-    cursor?: string | null
-    limit?: number
-}
-
-/**
- * Arguments for searching sections.
- */
-export type SearchSectionsArgs = SearchArgs & { projectId?: string | null }
-
-/**
- * Response from retrieving sections.
- * @see https://developer.todoist.com/api/v1/#tag/Sections/operation/get_sections_api_v1_sections_get
- */
-export type GetSectionsResponse = {
-    results: Section[]
-    nextCursor: string | null
-}
-
-/**
- * Arguments for creating a new section.
- * @see https://developer.todoist.com/api/v1/#tag/Sections/operation/create_section_api_v1_sections_post
- */
-export type AddSectionArgs = {
-    name: string
-    projectId: string
-    order?: number | null
-}
-
-/**
- * Arguments for updating a section.
- * @see https://developer.todoist.com/api/v1/#tag/Sections/operation/update_section_api_v1_sections__section_id__post
- */
-export type UpdateSectionArgs = {
-    name: string
-}
-
-/**
- * Arguments for retrieving labels.
- * @see https://developer.todoist.com/api/v1/#tag/Labels/operation/get_labels_api_v1_labels_get
- */
-export type GetLabelsArgs = {
-    cursor?: string | null
-    limit?: number
-}
-
-/**
- * Arguments for searching labels.
- */
-export type SearchLabelsArgs = SearchArgs
-
-/**
- * Response from retrieving labels.
- * @see https://developer.todoist.com/api/v1/#tag/Labels/operation/get_labels_api_v1_labels_get
- */
-export type GetLabelsResponse = {
-    results: Label[]
-    nextCursor: string | null
-}
-
-/**
- * Arguments for creating a new label.
- * @see https://developer.todoist.com/api/v1/#tag/Labels/operation/create_label_api_v1_labels_post
- */
-export type AddLabelArgs = {
-    name: string
-    order?: number | null
-    color?: ColorKey
-    isFavorite?: boolean
-}
-
-/**
- * Arguments for updating a label.
- * @see https://developer.todoist.com/api/v1/#tag/Labels/operation/update_label_api_v1_labels__label_id__post
- */
-export type UpdateLabelArgs = {
-    name?: string
-    order?: number | null
-    color?: ColorKey
-    isFavorite?: boolean
-}
-
-/**
- * Arguments for retrieving shared labels.
- * @see https://developer.todoist.com/api/v1/#tag/Labels/operation/shared_labels_api_v1_labels_shared_get
- */
-export type GetSharedLabelsArgs = {
-    omitPersonal?: boolean
-    cursor?: string | null
-    limit?: number
-}
-
-/**
- * Response from retrieving shared labels.
- * @see https://developer.todoist.com/api/v1/#tag/Labels/operation/shared_labels_api_v1_labels_shared_get
- */
-export type GetSharedLabelsResponse = {
-    results: string[]
-    nextCursor: string | null
-}
-
-/**
- * Arguments for renaming a shared label.
- * @see https://developer.todoist.com/api/v1/#tag/Labels/operation/shared_labels_rename_api_v1_labels_shared_rename_post
- */
-export type RenameSharedLabelArgs = {
-    name: string
-    newName: string
-}
-
-/**
- * Arguments for removing a shared label.
- * @see https://developer.todoist.com/api/v1/#tag/Labels/operation/shared_labels_remove_api_v1_labels_shared_remove_post
- */
-export type RemoveSharedLabelArgs = {
-    name: string
-}
-
-/**
- * Arguments for retrieving comments.
- * @see https://developer.todoist.com/api/v1/#tag/Comments/operation/get_comments_api_v1_comments_get
- */
-export type GetCommentsArgs = {
-    taskId: string
-    projectId?: never
-    cursor?: string | null
-    limit?: number
-}
-
-/**
- * Arguments for retrieving task comments.
- * @see https://developer.todoist.com/api/v1/#tag/Comments/operation/get_comments_api_v1_comments_get
- */
-export type GetTaskCommentsArgs = {
-    taskId: string
-    projectId?: never
-    cursor?: string | null
-    limit?: number
-}
-
-/**
- * Arguments for retrieving project comments.
- * @see https://developer.todoist.com/api/v1/#tag/Comments/operation/get_comments_api_v1_comments_get
- */
-export type GetProjectCommentsArgs = {
-    projectId: string
-    taskId?: never
-    cursor?: string | null
-    limit?: number
-}
-
-/**
- * Response from retrieving comments.
- * @see https://developer.todoist.com/api/v1/#tag/Comments/operation/get_comments_api_v1_comments_get
- */
-export type GetCommentsResponse = {
-    results: Comment[]
-    nextCursor: string | null
-}
-
-/**
- * Arguments for creating a new comment.
- * @see https://developer.todoist.com/api/v1/#tag/Comments/operation/create_comment_api_v1_comments_post
- */
-export type AddCommentArgs = {
-    content: string
-    attachment?: {
-        fileName?: string
-        fileUrl: string
-        fileType?: string
-        resourceType?: string
-    } | null
-} & RequireExactlyOne<{
-    taskId?: string
-    projectId?: string
-}>
-
-/**
- * Arguments for updating a comment.
- * @see https://developer.todoist.com/api/v1/#tag/Comments/operation/update_comment_api_v1_comments__comment_id__post
- */
-export type UpdateCommentArgs = {
-    content: string
-}
-
-/** Available reminder delivery services. */
-export const REMINDER_DELIVERY_SERVICES = ['email', 'push'] as const
-/** Delivery service for a reminder notification. */
-export type ReminderDeliveryService = (typeof REMINDER_DELIVERY_SERVICES)[number]
-export type ReminderDueDate = Partial<
-    Pick<DueDate, 'date' | 'string' | 'timezone' | 'lang' | 'isRecurring'>
->
-
-type ReminderTaskArgs = {
-    taskId: string
-}
-
-type TimeBasedReminderArgs = {
-    service?: ReminderDeliveryService
-    notifyUid?: string
-    isUrgent?: boolean
-}
-
-type AddLocationReminderFields = {
-    notifyUid?: string
-    name: string
-    locLat: string
-    locLong: string
-    locTrigger: LocationTrigger
-    radius?: number
-}
-
-export type AddRelativeReminderArgs = ReminderTaskArgs & {
-    reminderType?: 'relative'
-    minuteOffset: number
-} & TimeBasedReminderArgs
-
-export type AddAbsoluteReminderArgs = ReminderTaskArgs & {
-    reminderType: 'absolute'
-    due: ReminderDueDate
-} & TimeBasedReminderArgs
-
-export type AddLocationReminderArgs = ReminderTaskArgs & AddLocationReminderFields
-
-/**
- * Arguments for creating a new reminder.
- * @see https://developer.todoist.com/api/v1/#tag/Reminders/operation/create_reminder_api_v1_reminders_post
- */
-export type AddReminderArgs = AddRelativeReminderArgs | AddAbsoluteReminderArgs
-
-export type UpdateRelativeReminderArgs = {
-    reminderType: 'relative'
-} & Partial<Omit<AddRelativeReminderArgs, 'taskId' | 'reminderType'>>
-
-export type UpdateAbsoluteReminderArgs = {
-    reminderType: 'absolute'
-} & Partial<Omit<AddAbsoluteReminderArgs, 'taskId' | 'reminderType'>>
-
-/**
- * Arguments for updating an existing reminder.
- * @see https://developer.todoist.com/api/v1/#tag/Reminders/operation/update_reminder_api_v1_reminders__reminder_id__post
- */
-export type UpdateReminderArgs = UpdateRelativeReminderArgs | UpdateAbsoluteReminderArgs
-
-/**
- * Arguments for updating an existing location reminder.
- */
-export type UpdateLocationReminderArgs = Partial<Omit<AddLocationReminderArgs, 'taskId'>>
-
-/**
- * Common arguments for retrieving activity logs (excluding filter-type params).
- */
-type GetActivityLogsArgsCommon = {
-    /**
-     * Filter by the ID of a specific object.
-     */
-    objectId?: string
-    /**
-     * Filter events by parent project ID.
-     */
-    parentProjectId?: string
-    /**
-     * Filter events by parent task ID.
-     */
-    parentItemId?: string
-    /**
-     * When true, includes the parent object data in the response.
-     */
-    includeParentObject?: boolean
-    /**
-     * When true, includes child object data in the response.
-     */
-    includeChildObjects?: boolean
-    /**
-     * Filter by the user ID who initiated the event.
-     */
-    initiatorId?: string
-    /**
-     * When true, filters for events with no initiator (system-generated events).
-     * When false, filters for events with an initiator.
-     * When null or undefined, no filtering on initiator is applied.
-     */
-    initiatorIdNull?: boolean | null
-    /**
-     * When true, ensures the last state of objects is included in the response.
-     */
-    ensureLastState?: boolean
-    /**
-     * When true, includes comment annotations in the response.
-     */
-    annotateNotes?: boolean
-    /**
-     * When true, includes parent object annotations in the response.
-     */
-    annotateParents?: boolean
-    /**
-     * Pagination cursor for retrieving the next page of results.
-     */
-    cursor?: string | null
-    /**
-     * Maximum number of results to return per page.
-     */
-    limit?: number
-    /**
-     * Start date for filtering events (inclusive). Replaces deprecated `since`.
-     * Accepts a Date object or YYYY-MM-DD string.
-     */
-    dateFrom?: Date | string
-    /**
-     * End date for filtering events (inclusive). Replaces deprecated `until`.
-     * Accepts a Date object or YYYY-MM-DD string.
-     */
-    dateTo?: Date | string
-}
-
-export type GetActivityLogsArgs = GetActivityLogsArgsCommon & {
-    objectEventTypes?: ActivityObjectEventType | ActivityObjectEventType[]
-    /** @removed Use `objectEventTypes` instead. */
-    objectType?: never
-    /** @removed Use `objectEventTypes` instead. */
-    eventType?: never
-    /** @removed Use `dateFrom` instead. */
-    since?: never
-    /** @removed Use `dateTo` instead. */
-    until?: never
-}
-
-/**
- * Response from retrieving activity logs.
- */
-export type GetActivityLogsResponse = {
-    results: ActivityEvent[]
-    nextCursor: string | null
-}
-
-/**
- * Arguments for uploading a file.
- * @see https://developer.todoist.com/api/v1/#tag/Uploads/operation/upload_file_api_v1_uploads_post
- */
-export type UploadFileArgs = {
-    /**
-     * The file content to upload. Can be:
-     * - Buffer: File content as a Buffer (requires fileName)
-     * - ReadableStream: File content as a stream (requires fileName)
-     * - string: Path to a file on the filesystem (fileName is optional, will be inferred from path)
-     */
-    file: Buffer | NodeJS.ReadableStream | string | Blob
-    /**
-     * The name of the file. Required for Buffer and Stream inputs.
-     * Optional for file path strings and File objects (will be inferred if not provided).
-     */
-    fileName?: string
-    /**
-     * The project ID to associate the upload with.
-     */
-    projectId?: string | null
-}
-
-/**
- * Arguments for deleting an uploaded file.
- * @see https://developer.todoist.com/api/v1/#tag/Uploads/operation/delete_upload_api_v1_uploads_delete
- */
-export type DeleteUploadArgs = {
-    /**
-     * The URL of the file to delete.
-     */
-    fileUrl: string
-}
-
-/**
- * Arguments for moving a project to a workspace.
- */
-export type MoveProjectToWorkspaceArgs = {
-    /** The ID of the project to move. */
-    projectId: string
-    /** The target workspace ID. */
-    workspaceId: string
-    /** Optional target folder ID within the workspace. */
-    folderId?: string | null
-    /** Optional access settings for the project in the workspace. */
-    access?: {
-        visibility: ProjectVisibility
-    }
-}
-
-/**
- * Arguments for moving a project to personal.
- */
-export type MoveProjectToPersonalArgs = {
-    /** The ID of the project to move. */
-    projectId: string
-}
-
-// Project extras
-
-/**
- * Arguments for counting archived projects.
- * @see https://developer.todoist.com/api/v1/#tag/Projects/operation/count_projects_archived_api_v1_projects_archived_count_get
- */
-export type GetArchivedProjectsCountArgs = {
-    workspaceId?: number | null
-    joined?: boolean | null
-}
-
-/**
- * Response from counting archived projects.
- * @see https://developer.todoist.com/api/v1/#tag/Projects/operation/count_projects_archived_api_v1_projects_archived_count_get
- */
-export type GetArchivedProjectsCountResponse = {
-    count: number
-}
-
-/**
- * An action permitted for a role.
- */
-export type ActionView = {
-    name: string
-}
-
-/**
- * A project collaborator role with its permitted actions.
- */
-export type ProjectRoleView = {
-    name: CollaboratorRole
-    actions: ActionView[]
-}
-
-/**
- * A workspace role with its permitted actions.
- */
-export type WorkspaceRoleView = {
-    name: WorkspaceRole
-    actions: ActionView[]
-}
-
-/**
- * Response from getting project permissions (role-to-action mappings).
- * @see https://developer.todoist.com/api/v1/#tag/Projects/operation/permissions_api_v1_projects_permissions_get
- */
-export type GetProjectPermissionsResponse = {
-    projectCollaboratorActions: ProjectRoleView[]
-    workspaceCollaboratorActions: WorkspaceRoleView[]
-}
-
-/**
- * Arguments for getting full project data.
- * @see https://developer.todoist.com/api/v1/#tag/Projects/operation/projects_full_data_api_v1_projects__project_id__full_get
- */
-export type GetFullProjectArgs = {
-    /**
-     * Required to access the public project without authentication.
-     */
-    publicKey?: string | null
-}
-
-/**
- * Response from getting full project data.
- * @see https://developer.todoist.com/api/v1/#tag/Projects/operation/projects_full_data_api_v1_projects__project_id__full_get
- */
-export type GetFullProjectResponse = {
-    project: (PersonalProject | WorkspaceProject) | null
-    commentsCount: number
-    tasks: Task[]
-    sections: Section[]
-    collaborators: User[]
-    notes: Comment[]
-}
-
-// Backup types
-
-/**
- * Arguments for listing backups.
- * @see https://developer.todoist.com/api/v1/#tag/Backups/operation/get_backups_api_v1_backups_get
- */
-export type GetBackupsArgs = {
-    /** MFA token if required. Not needed when using an OAuth token with the `backups:read` scope. */
-    mfaToken?: string | null
-}
-
-/**
- * Arguments for downloading a backup.
- * @see https://developer.todoist.com/api/v1/#tag/Backups/operation/download_backup_api_v1_backups_download_get
- */
-export type DownloadBackupArgs = {
-    /** The backup file URL. */
-    file: string
-}
-
-// Email forwarding types
-
-/** Object types that support email forwarding. */
-export const EMAIL_OBJECT_TYPES = ['project', 'project_comments', 'task'] as const
-/** Object type that supports email forwarding. */
-export type EmailObjectType = (typeof EMAIL_OBJECT_TYPES)[number]
-
-/**
- * Arguments for getting or creating an email forwarding address.
- * @see https://developer.todoist.com/api/v1/#tag/Emails/operation/email_get_or_create_api_v1_emails_put
- */
-export type GetOrCreateEmailArgs = {
-    /** The type of object to forward emails to. */
-    objType: EmailObjectType
-    /** The ID of the object. */
-    objId: string
-}
-
-/**
- * Response from getting or creating an email forwarding address.
- */
-export type GetOrCreateEmailResponse = {
-    email: string
-}
-
-/**
- * Arguments for disabling email forwarding.
- * @see https://developer.todoist.com/api/v1/#tag/Emails/operation/email_disable_api_v1_emails_delete
- */
-export type DisableEmailArgs = {
-    /** The type of object to disable forwarding for. */
-    objType: EmailObjectType
-    /** The ID of the object. */
-    objId: string
-}
-
-// ID mapping types
-
-/** Object types that support ID mappings. */
-export const ID_MAPPING_OBJECT_TYPES = [
-    'sections',
-    'tasks',
-    'comments',
-    'reminders',
-    'location_reminders',
-    'projects',
-] as const
-/** Object type that supports ID mappings. */
-export type IdMappingObjectType = (typeof ID_MAPPING_OBJECT_TYPES)[number]
-
-/** Object types that support moved ID lookups. */
-export const MOVED_ID_OBJECT_TYPES = [
-    'sections',
-    'tasks',
-    'comments',
-    'reminders',
-    'location_reminders',
-] as const
-/** Object type that supports moved ID lookups. */
-export type MovedIdObjectType = (typeof MOVED_ID_OBJECT_TYPES)[number]
-
-/**
- * Arguments for getting ID mappings between old and new IDs.
- * @see https://developer.todoist.com/api/v1/#tag/Ids/operation/id_mappings_api_v1_id_mappings__obj_name___obj_ids__get
- */
-export type GetIdMappingsArgs = {
-    /** The type of object to look up. */
-    objName: IdMappingObjectType
-    /** Array of IDs to look up. */
-    objIds: [string, ...string[]]
-}
-
-/**
- * Arguments for getting moved IDs.
- * @see https://developer.todoist.com/api/v1/#tag/Ids/operation/moved_ids_api_v1_moved_ids__obj_name__get
- */
-export type GetMovedIdsArgs = {
-    /** The type of object to look up. */
-    objName: MovedIdObjectType
-    /** Array of old IDs to look up. */
-    oldIds?: string[]
-}
-
-// Reminders list types
-
-/**
- * Arguments for listing reminders.
- * @see https://developer.todoist.com/api/v1/#tag/Reminders/operation/get_reminders_api_v1_reminders_get
- */
-export type GetRemindersArgs = {
-    /** Filter by task ID. */
-    taskId?: string | null
-    /** Cursor for pagination. */
-    cursor?: string | null
-    /** Number of results per page (max 200, default 50). */
-    limit?: number
-}
-
-/**
- * Paginated response for reminders.
- */
-export type GetRemindersResponse = {
-    results: Reminder[]
-    nextCursor: string | null
-}
-
-/**
- * Arguments for listing location reminders.
- * @see https://developer.todoist.com/api/v1/#tag/Location-reminders/operation/get_location_reminders_api_v1_location_reminders_get
- */
-export type GetLocationRemindersArgs = {
-    /** Filter by task ID. */
-    taskId?: string | null
-    /** Cursor for pagination. */
-    cursor?: string | null
-    /** Number of results per page (max 200, default 50). */
-    limit?: number
-}
-
-/**
- * Paginated response for location reminders.
- */
-export type GetLocationRemindersResponse = {
-    results: LocationReminder[]
-    nextCursor: string | null
-}
-
-// Completed tasks (legacy) types
-
-/**
- * Arguments for getting all completed tasks.
- * @see https://developer.todoist.com/api/v1/#tag/Tasks/operation/get_all_completed_items_api_v1_tasks_completed_get
- */
-export type GetAllCompletedTasksArgs = {
-    /** Filter by project ID. */
-    projectId?: string | null
-    /** Filter by label name. */
-    label?: string | null
-    /** Number of results to return (max 200, default 30). */
-    limit?: number
-    /** Number of results to skip (default 0). */
-    offset?: number
-    /** Return items completed after this date. */
-    since?: Date | null
-    /** Return items completed before this date. */
-    until?: Date | null
-    /** Include comment data in the response. */
-    annotateNotes?: boolean
-    /** Include task data in the response. */
-    annotateItems?: boolean
-}
-
-/**
- * Response from getting all completed tasks.
- */
-export type GetAllCompletedTasksResponse = {
-    projects: Record<string, Record<string, unknown>>
-    sections: Record<string, Record<string, unknown>>
-    items: Task[]
-}
-
-// Template types
-
-/**
- * Arguments for exporting a project as a template file.
- * @see https://developer.todoist.com/api/v1/#tag/Templates/operation/export_as_file_api_v1_templates_file_get
- */
-export type ExportTemplateFileArgs = {
-    /** The project ID to export. */
-    projectId: string
-    /** Whether to use relative dates in the export. */
-    useRelativeDates?: boolean
-}
-
-/**
- * Arguments for exporting a project as a template URL.
- * @see https://developer.todoist.com/api/v1/#tag/Templates/operation/export_as_url_api_v1_templates_url_get
- */
-export type ExportTemplateUrlArgs = {
-    /** The project ID to export. */
-    projectId: string
-    /** Whether to use relative dates in the export. */
-    useRelativeDates?: boolean
-}
-
-/**
- * Response from exporting a project as a template URL.
- */
-export type ExportTemplateUrlResponse = {
-    fileName: string
-    fileUrl: string
-}
-
-/**
- * Arguments for creating a project from a template file.
- * @see https://developer.todoist.com/api/v1/#tag/Templates/operation/create_project_from_file_api_v1_templates_create_project_from_file_post
- */
-export type CreateProjectFromTemplateArgs = {
-    /** Name for the new project. */
-    name: string
-    /** The template file content. */
-    file: Buffer | NodeJS.ReadableStream | string | Blob
-    /** Optional file name (required for Buffer/Stream). */
-    fileName?: string
-    /** Optional workspace ID. */
-    workspaceId?: string | null
-}
-
-/**
- * Response from creating a project from a template.
- */
-export type CreateProjectFromTemplateResponse = {
-    status: string
-    projectId: string
-    templateType: string
-    projects: (PersonalProject | WorkspaceProject)[]
-    sections: Section[]
-    tasks: Task[]
-    comments: Comment[]
-}
-
-/**
- * Arguments for importing a template file into an existing project.
- * @see https://developer.todoist.com/api/v1/#tag/Templates/operation/import_into_project_from_file_api_v1_templates_import_into_project_from_file_post
- */
-export type ImportTemplateIntoProjectArgs = {
-    /** The project ID to import into. */
-    projectId: string
-    /** The template file content. */
-    file: Buffer | NodeJS.ReadableStream | string | Blob
-    /** Optional file name (required for Buffer/Stream). */
-    fileName?: string
-}
-
-/**
- * Arguments for importing a template by ID into an existing project.
- * @see https://developer.todoist.com/api/v1/#tag/Templates/operation/import_into_project_from_template_id_api_v1_templates_import_into_project_from_template_id_post
- */
-export type ImportTemplateFromIdArgs = {
-    /** The project ID to import into. */
-    projectId: string
-    /** The template ID to import. */
-    templateId: string
-    /** Locale for the import (default: 'en'). */
-    locale?: string
-}
-
-/**
- * Response from importing a template into a project.
- */
-export type ImportTemplateResponse = {
-    status: string
-    templateType: string
-    projects: (PersonalProject | WorkspaceProject)[]
-    sections: Section[]
-    tasks: Task[]
-    comments: Comment[]
-}
-
-// Insights types
-
-/** Known activity object types. Accepts any string for forward compatibility. */
-export type InsightsObjectType = 'ITEM' | 'PROJECT' | 'NOTE' | (string & Record<string, never>)
-
-/** Known activity event types. Accepts any string for forward compatibility. */
-export type InsightsEventType =
-    | 'ADDED'
-    | 'DELETED'
-    | 'UPDATED'
-    | 'ARCHIVED'
-    | 'UNARCHIVED'
-    | 'COMPLETED'
-    | 'UNCOMPLETED'
-    | 'SHARED'
-    | 'LEFT'
-    | (string & Record<string, never>)
-
-/**
- * Arguments for getting project activity stats.
- * @see https://developer.todoist.com/api/v1/#tag/Insights/operation/get_project_activity_stats_api_v1_projects__project_id__insights_activity_stats_get
- */
-export type GetProjectActivityStatsArgs = {
-    /** The type of object to get activity counts for (default: 'ITEM'). */
-    objectType?: InsightsObjectType
-    /** The type of event to count (default: 'COMPLETED'). */
-    eventType?: InsightsEventType
-    /** Number of weeks of activity counts to retrieve (1-12, default 2). */
-    weeks?: number
-    /** Whether to include weekly rollup counts in the response. */
-    includeWeeklyCounts?: boolean
-}
-
-/**
- * Arguments for getting workspace insights.
- * @see https://developer.todoist.com/api/v1/#tag/Insights/operation/get_workspace_insights_api_v1_workspaces__workspace_id__insights_get
- */
-export type GetWorkspaceInsightsArgs = {
-    /** Project IDs to get insights for. */
-    projectIds?: string[]
-}
-
-// Workspace-related types
-
-/**
- * Arguments for creating a new workspace.
- * @see https://developer.todoist.com/api/v1/#tag/Workspace/operation/create_workspace_api_v1_workspaces_post
- */
-export type AddWorkspaceArgs = {
-    /** Name of the workspace. */
-    name: string
-    /** Description of the workspace. */
-    description?: string | null
-    /** Whether link sharing is enabled for the workspace. */
-    isLinkSharingEnabled?: boolean
-    /** Whether guests are allowed in the workspace. */
-    isGuestAllowed?: boolean
-    /** Workspace email domain. */
-    domainName?: string | null
-    /** Whether users with matching email domains can discover this workspace. */
-    domainDiscovery?: boolean
-    /** Whether workspace invites are restricted to workspace domain. */
-    restrictEmailDomains?: boolean
-    /** Workspace properties. */
-    properties?: Record<string, unknown> | null
-}
-
-/**
- * Arguments for updating an existing workspace.
- * @see https://developer.todoist.com/api/v1/#tag/Workspace/operation/update_workspace_api_v1_workspaces__workspace_id__post
- */
-export type UpdateWorkspaceArgs = {
-    /** Updated workspace name. */
-    name?: string
-    /** Updated workspace description. */
-    description?: string | null
-    /** Updated link sharing status. */
-    isLinkSharingEnabled?: boolean
-    /** Updated guest access status. */
-    isGuestAllowed?: boolean | null
-    /** Updated workspace email domain. */
-    domainName?: string | null
-    /** Updated domain discovery setting. */
-    domainDiscovery?: boolean
-    /** Updated email domain restriction setting. */
-    restrictEmailDomains?: boolean
-    /** Updated workspace properties. */
-    properties?: Record<string, unknown> | null
-    /** Updated collapse state for current user. */
-    isCollapsed?: boolean
-}
-
-/**
- * Arguments for getting workspace members activity.
- * @see https://developer.todoist.com/api/v1/#tag/Workspace/operation/get_workspace_members_activity_api_v1_workspaces_members_get
- */
-export type GetWorkspaceMembersActivityArgs = {
-    /** The workspace ID. */
-    workspaceId: string
-    /** Comma-separated list of user IDs to filter by. */
-    userIds?: string | null
-    /** Comma-separated list of project IDs to filter by. */
-    projectIds?: string | null
-}
-
-/**
- * Response from getting workspace members activity.
- */
-export type GetWorkspaceMembersActivityResponse = {
-    members: MemberActivityInfo[]
-}
-
-/**
- * Arguments for getting tasks assigned to a workspace user.
- * @see https://developer.todoist.com/api/v1/#tag/Workspace/operation/get_workspace_user_tasks_api_v1_workspaces__workspace_id__users__user_id__tasks_get
- */
-export type GetWorkspaceUserTasksArgs = {
-    /** The workspace ID. */
-    workspaceId: string
-    /** The user ID. */
-    userId: string
-    /** Comma-separated list of project IDs to filter by. */
-    projectIds?: string | null
-}
-
-/**
- * Response from getting workspace user tasks.
- */
-export type GetWorkspaceUserTasksResponse = {
-    tasks: WorkspaceUserTask[]
-}
-
-/**
- * Arguments for inviting users to a workspace.
- * @see https://developer.todoist.com/api/v1/#tag/Workspace/operation/invite_workspace_users_api_v1_workspaces__workspace_id__users_invite_post
- */
-export type InviteWorkspaceUsersArgs = {
-    /** The workspace ID. */
-    workspaceId: string
-    /** List of user emails to invite. */
-    emailList: string[]
-    /** Role assigned to invited users. */
-    role?: WorkspaceRole
-}
-
-/**
- * Response from inviting workspace users.
- */
-export type InviteWorkspaceUsersResponse = {
-    invitedEmails: string[]
-}
-
-/**
- * Arguments for updating a workspace user's role.
- * @see https://developer.todoist.com/api/v1/#tag/Workspace/operation/update_workspace_user_api_v1_workspaces__workspace_id__users__user_id__post
- */
-export type UpdateWorkspaceUserArgs = {
-    /** The workspace ID. */
-    workspaceId: string
-    /** The user ID. */
-    userId: string
-    /** Updated role for the user. */
-    role: WorkspaceRole
-}
-
-/**
- * Arguments for removing a user from a workspace.
- * @see https://developer.todoist.com/api/v1/#tag/Workspace/operation/remove_workspace_user_api_v1_workspaces__workspace_id__users__user_id__delete
- */
-export type RemoveWorkspaceUserArgs = {
-    /** The workspace ID. */
-    workspaceId: string
-    /** The user ID. */
-    userId: string
-}
-
-/**
- * Arguments for getting workspace invitations.
- */
-export type GetWorkspaceInvitationsArgs = {
-    /**
-     * The workspace ID to get invitations for.
-     */
-    workspaceId: number
-}
-
-/**
- * Arguments for deleting a workspace invitation.
- */
-export type DeleteWorkspaceInvitationArgs = {
-    /**
-     * The workspace ID.
-     */
-    workspaceId: number
-    /**
-     * The email address of the invitation to delete.
-     */
-    userEmail: string
-}
-
-/**
- * Arguments for accepting/rejecting a workspace invitation.
- */
-export type WorkspaceInvitationActionArgs = {
-    /**
-     * The invitation code from the email.
-     */
-    inviteCode: string
-}
-
-/**
- * Arguments for joining a workspace.
- */
-export type JoinWorkspaceArgs = {
-    /**
-     * Optional invitation code/link to join via.
-     */
-    inviteCode?: string | null
-    /**
-     * Optional workspace ID to join via domain auto-join.
-     */
-    workspaceId?: number | null
-}
-
-/**
- * Arguments for uploading/updating workspace logo.
- */
-export type WorkspaceLogoArgs = {
-    /**
-     * The workspace ID.
-     */
-    workspaceId: number
-    /**
-     * The image file to upload (Buffer, Stream, or file path).
-     */
-    file?: Buffer | NodeJS.ReadableStream | string | Blob
-    /**
-     * The file name (required for Buffer/Stream uploads).
-     */
-    fileName?: string
-    /**
-     * Whether to delete the logo instead of updating it.
-     */
-    delete?: boolean
-}
-
-/**
- * Arguments for getting workspace plan details.
- */
-export type GetWorkspacePlanDetailsArgs = {
-    /**
-     * The workspace ID.
-     */
-    workspaceId: number
-}
-
-/**
- * Arguments for getting workspace users (paginated).
- */
-export type GetWorkspaceUsersArgs = {
-    /**
-     * Optional workspace ID. If not provided, returns users for all workspaces.
-     */
-    workspaceId?: number | null
-    /**
-     * Cursor for pagination.
-     */
-    cursor?: string | null
-    /**
-     * Maximum number of users to return (default: 100).
-     */
-    limit?: number
-}
-
-/**
- * Arguments for getting workspace projects (paginated).
- */
-export type GetWorkspaceProjectsArgs = {
-    /**
-     * The workspace ID.
-     */
-    workspaceId: number
-    /**
-     * Cursor for pagination.
-     */
-    cursor?: string | null
-    /**
-     * Maximum number of projects to return (default: 100).
-     */
-    limit?: number
-}
-
-/**
- * Paginated response for workspace users.
- */
-export type GetWorkspaceUsersResponse = {
-    /**
-     * Whether there are more users available.
-     */
-    hasMore: boolean
-    /**
-     * Cursor for the next page of results.
-     */
-    nextCursor?: string
-    /**
-     * Array of workspace users.
-     */
-    workspaceUsers: WorkspaceUser[]
-}
-
-/**
- * Response type for workspace invitations endpoint (simple email list).
- */
-export type WorkspaceInvitationsResponse = string[]
-
-/**
- * Response type for all workspace invitations endpoint (detailed objects).
- */
-export type AllWorkspaceInvitationsResponse = WorkspaceInvitation[]
-
-/**
- * Response type for workspace logo upload.
- */
-export type WorkspaceLogoResponse = Record<string, unknown> | null
