+++ src/utils/validators.ts
@@ -1,19 +1,19 @@
 import { type ZodType } from 'zod'
 
+import { TaskSchema } from '../types/tasks/types'
 import {
-    AttachmentSchema,
-    SectionSchema,
-    LabelSchema,
-    CommentSchema,
-    UserSchema,
-    CurrentUserSchema,
-    TaskSchema,
     PersonalProjectSchema,
     WorkspaceProjectSchema,
     type WorkspaceProject,
     type PersonalProject,
-    ProductivityStatsSchema,
-    ActivityEventSchema,
+} from '../types/projects/types'
+import { SectionSchema } from '../types/sections/types'
+import { LabelSchema } from '../types/labels/types'
+import { AttachmentSchema, CommentSchema } from '../types/comments/types'
+import { UserSchema, CurrentUserSchema } from '../types/users/types'
+import { ProductivityStatsSchema } from '../types/productivity/types'
+import { ActivityEventSchema } from '../types/activity/types'
+import {
     WorkspaceUserSchema,
     type WorkspaceUser,
     WorkspaceInvitationSchema,
@@ -22,15 +22,16 @@ import {
     WorkspaceSchema,
     MemberActivityInfoSchema,
     WorkspaceUserTaskSchema,
+} from '../types/workspaces/types'
+import {
     ProjectActivityStatsSchema,
     ProjectHealthSchema,
     ProjectHealthContextSchema,
     ProjectProgressSchema,
     WorkspaceInsightsSchema,
-    BackupSchema,
-    IdMappingSchema,
-    MovedIdSchema,
-} from '../types/entities'
+} from '../types/insights/types'
+import { BackupSchema } from '../types/backups/types'
+import { IdMappingSchema, MovedIdSchema } from '../types/id-mappings/types'
 
 import {
     FilterSchema,
